[type:feat] Make an Apache Shenyu SSO authentication plugin based
