+++ shenyu-bootstrap/pom.xml
@@ -157,7 +157,7 @@
             <artifactId>shenyu-spring-boot-starter-plugin-auth</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <!--shenyu jwt end-->
+        <!--shenyu auth plugin end-->
         <!-- shenyu modify response plugin start-->
         <dependency>
             <groupId>org.apache.shenyu</groupId>
