--- shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-sprimg-boot-starter-plugin-auth/src/main/java/org/apache/shenyu/springboot/starter/plugin/auth/AuthPluginConfiguration.java
@@ -1,30 +1,0 @@
-// Copyright 2022 The casdoor Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-package org.apache.shenyu.springboot.starter.plugin.auth;
-
-import org.apache.shenyu.plugin.api.ShenyuPlugin;
-import org.apache.shenyu.plugin.auth.AuthPlugin;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConditionalOnProperty(value = {"shenyu.plugins.auth.enabled"}, havingValue = "true", matchIfMissing = true)
-public class AuthPluginConfiguration {
-
-    @Bean
-    public ShenyuPlugin contextPathPlugin() {
-        return new AuthPlugin();
-    }
-}
