--- shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-sprimg-boot-starter-plugin-auth/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.springboot.starter.plugin.auth.AuthPluginConfiguration
