--- shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-sprimg-boot-starter-plugin-auth/src/main/resources/META-INF/spring.provides
@@ -1,1 +1,0 @@
-provides: shenyu-spring-boot-starter-plugin-auth
